#126 move szinterface to sz
