+++ sz/flags.go
@@ -1,4 +1,4 @@
-package szinterface
+package sz
 
 // ----------------------------------------------------------------------------
 // Constants
