+++ sz/parameters.go
@@ -1,4 +1,4 @@
-package szinterface
+package sz
 
 // ----------------------------------------------------------------------------
 // Constants
@@ -8,7 +8,9 @@ const (
 	SZ_INITIALIZE_WITH_DEFAULT_CONFIGURATION int64  = 0
 	SZ_NO_ATTRIBUTES                         string = ""
 	SZ_NO_EXCLUSIONS                         string = ""
+	SZ_NO_LOGGING                            int64  = 0
 	SZ_NO_REQUIRED_DATASOURCES               string = ""
 	SZ_NO_SEARCH_PROFILE                     string = ""
+	SZ_VERBOSE_LOGGING                       int64  = 1
 	SZ_WITHOUT_INFO                          int64  = 0
 )
