+++ sz/szi_test.go
@@ -1,4 +1,4 @@
-package szinterface
+package sz
 
 import (
 	"testing"
